rename files to `.schema.json`
