+++ packages/json_schemas/schemas/key_value_store.schema.json
