+++ packages/json_schemas/src/actor.schema.ts
@@ -1,7 +1,7 @@
 import { ACTOR_LIMITS } from '@apify/consts';
 
-// NOTE: When updating actor.json schema, always update this file not the schemas/actor.json file directly.
-// The schemas/actor.json file is generated from this file during the build step.
+// NOTE: When updating actor schema, always update this file not the schemas/actor.schema.json file directly.
+// The schemas/actor.schema.json file is generated from this file during the build step.
 
 export const actorSchema = {
     $id: 'actor.json',
