+++ packages/json_schemas/src/schemas.ts
@@ -1,8 +1,8 @@
-import actorSchema from '../schemas/actor.json';
-import datasetSchema from '../schemas/dataset.json';
-import inputSchema from '../schemas/input.json';
-import keyValueStoreSchema from '../schemas/key_value_store.json';
-import outputSchema from '../schemas/output.json';
+import actorSchema from '../schemas/actor.schema.json';
+import datasetSchema from '../schemas/dataset.schema.json';
+import inputSchema from '../schemas/input.schema.json';
+import keyValueStoreSchema from '../schemas/key_value_store.schema.json';
+import outputSchema from '../schemas/output.schema.json';
 
 export {
     actorSchema,
